Performance Improvement.
All vectors are now pre-allocated once, unless the input size changes. This should not happen if the object is used within the pfft~ patcher environment. That said, if user for some reason uses the object in the regular audio patching environment, the size can be changed mid operation if the user changes Max's audio buffer size.