removed special case treatments for passthrough (alpha zero) and reversal ±2.
The torch implementation just returned the input or the reverse directly without applying the algo, here we run the algo for this.

That said, once going through the algo at the mid point of the signals there is a discrepency which we fix by simply replacing this sample with the average of the neighbouring samples.